\name{as.RV}
\alias{as.RV}
\title{Turn a probability vector with possible outcome values in the 'names()' attribute
into a random variable:}
\usage{
as.RV(px)
}
\arguments{
  \item{px}{A probability vector with possible outcome
  values in the 'names()' attribute}
}
\description{
Turn a probability vector with possible outcome values in
the 'names()' attribute into a random variable:
}

