% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{iid}
\alias{iid}
\title{Probability mass function of  X^n}
\usage{
iid(X, n = 2, sep = ",", fractions = attr(X, "fractions"))
}
\arguments{
\item{X}{random variable}

\item{n}{power}

\item{sep}{separator between items from marginal distributions, by default set to ","}

\item{fractions}{If TRUE, return the probabilities as fractions}
}
\description{
Probability mass function of  X^n
}
\examples{
d <- RV(c("A","B","C"), odds = c(3,5,11))
d2 <- iid(d)
probs(d2)
}
\author{
Heike Hofmann \email{hofmann@iastate.edu}
}
