% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{independent}
\alias{independent}
\title{Tests whether the random variables X and Y are independent}
\usage{
independent(X, Y)
}
\arguments{
\item{X}{A random variable}

\item{Y}{A random variable}
}
\description{
Tests whether the random variables X and Y are independent
}
\examples{
AandB <- jointRV(outcomes = list(1:3, 0:2), probs = 1:9 / sum(1:9))
A <- marginal(AandB, 1)
B <- marginal(AandB, 2)
independent(A, B) # FALSE
CandD <- jointRV(outcomes = list(1:3, 0:2))
C <- marginal(CandD, 1)
D <- marginal(CandD, 2)
independent(C, D) # FALSE
}
\author{
Eric Hare \email{erichare@iastate.edu}
}

