% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{joint}
\alias{joint}
\title{Joint probability mass function of random variables X and Y}
\usage{
joint(X, Y, sep = ",", fractions = (attr(X, "fractions") & attr(Y,
  "fractions")))
}
\arguments{
\item{X}{random variable}

\item{Y}{random variable}

\item{sep}{separator between items from marginal distributions, by default set to ","}

\item{fractions}{If TRUE, return the probabilities as fractions}
}
\description{
Joint probability mass function of random variables X and Y
}
\examples{
d <- RV(c("A","B","C"), odds = c(3,5,11))
d2 <- joint(d,d)
probs(d2)
}
\author{
Heike Hofmann \email{hofmann@iastate.edu}
}

