% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{jointRV}
\alias{jointRV}
\title{Make a joint random variable consisting}
\usage{
jointRV(outcomes, probs = NULL, ...)
}
\arguments{
\item{outcomes}{The possible outcomes of the joint random variable, as a list}

\item{probs}{The probabilities of each event, in the order (x1, y1, x1, y2, ..., x2, y1, x2, y2, ..., xn, yn)}

\item{...}{Further arguments to be passed to the RV function}
}
\value{
An RV object
}
\description{
Make a joint random variable consisting
}

