% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{KURT}
\alias{KURT}
\title{Kurtosis of a random variable}
\usage{
KURT(X)
}
\arguments{
\item{X}{random variable}
}
\description{
Kurtosis of a random variable
}
\examples{
X.Bern <- make.RV(c(1,0), c(.5,.5))
KURT(X.Bern)
}

