% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{marginal}
\alias{marginal}
\title{Marginal distribution of a joint random variable}
\usage{
marginal(X, num)
}
\arguments{
\item{X}{A random variable}

\item{num}{Number indicating which marginal distribution to extract}
}
\description{
Extracts the marginal probability mass functions from a joint distribution.
}
\examples{
X <- RV(1:6, 1/6)
X3 <- iid(X, 3)
margins(X3)
}
\author{
Eric Hare \email{erichare@iastate.edu}
}

