% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multN}
\alias{multN}
\title{Probability mass function of  X^n}
\usage{
multN(X, n = 2, sep = ",", fractions = attr(X, "fractions"))
}
\arguments{
\item{X}{random variable}

\item{n}{power}

\item{sep}{separator between items from marginal distributions, by default set to ","}

\item{fractions}{If TRUE, return the probabilities as fractions}
}
\description{
Probability mass function of  X^n
}
\examples{
d <- make.RV(c("A","B","C"), odds = c(3,5,11))
d2 <- multN(d)
probs(d2)
}
\author{
Heike Hofmann \email{hofmann@iastate.edu}
}

