% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Prop}
\alias{Prop}
\title{Proportion of an event observed in a vector of simulated trials}
\usage{
Prop(X.sim)
}
\arguments{
\item{X.sim}{A simulated data vector produced with the 'rsim()' function}
}
\description{
Proportion of an event observed in a vector of simulated trials
}
\examples{
X <- make.RV(c(100000,10000,0), c(0.00025,0.005,0.99475))
X.sim <- rsim(200000, X)

Prop(X.sim>0)
Prop(X.sim==100000)
Prop(X.sim==2000)
}

