% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{SD}
\alias{SD}
\title{Standard deviation of a random variable}
\usage{
SD(X)
}
\arguments{
\item{X}{random variable}
}
\description{
Standard deviation of a random variable
}
\examples{
X.Bern <- RV(c(1,0), c(.5,.5))
E(X.Bern)
}

