% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{SKEW}
\alias{SKEW}
\title{Skewness of a random variable}
\usage{
SKEW(X)
}
\arguments{
\item{X}{random variable}
}
\description{
Skewness of a random variable
}
\examples{
X.Bern <- RV(c(1,0), c(.5,.5))
SKEW(X.Bern)
}
