% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{SofIID}
\alias{SofIID}
\title{Sum of independent identically distributed random variables}
\usage{
SofIID(X, n = 2, progress = TRUE, fractions = attr(X, "fractions"))
}
\arguments{
\item{X}{A random variable}

\item{n}{The number of Xs to sum}

\item{progress}{If TRUE, display a progress bar}

\item{fractions}{If TRUE, return the probabilities as fractions}
}
\description{
Sum of independent identically distributed random variables
}
\examples{
X.Bern <- make.RV(c(1,0), c(.5,.5))

S5 <- SofIID(X.Bern, 5)
S128 <- SofIID(X.Bern, 128)
}

