% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{V}
\alias{V}
\title{Variance of a random variable}
\usage{
V(X)
}
\arguments{
\item{X}{random variable}
}
\description{
Variance of a random variable
}
\examples{
X.Bern <- RV(c(1,0), c(.5,.5))
E(X.Bern)
}

