% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.RV}
\alias{as.RV}
\title{Turn a probability vector with possible outcome values in the 'names()' attribute
into a random variable:}
\usage{
as.RV(px, fractions = TRUE)
}
\arguments{
\item{px}{A probability vector with possible outcome values in the 'names()' attribute}

\item{fractions}{If TRUE, return the probabilities as fractions}
}
\description{
Turn a probability vector with possible outcome values in the 'names()' attribute
into a random variable:
}

