% Generated by roxygen2 (4.0.1): do not edit by hand
\name{E}
\alias{E}
\title{Expected value of a random variable}
\usage{
E(X)
}
\arguments{
\item{X}{random variable}
}
\description{
Expected value of a random variable
}
\examples{
X.Bern <- make.RV(c(1,0), c(.5,.5))
E(X.Bern)

X.fair.die <- make.RV(1:6, rep(1/6,6))
E(X.fair.die)
}

