% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{marginal}
\alias{marginal}
\title{Marginal distribution of a joint random variable}
\usage{
marginal(X, num)
}
\arguments{
\item{X}{A random variable}

\item{num}{Number indicating which marginal distribution to extract}
}
\description{
Extracts the marginal probability mass functions from a joint distribution.
}
\examples{
AandB <- jointRV(outcomes = list(1:3, 0:2), probs = 1:9 / sum(1:9))
marginal(AandB, 1)
marginal(AandB, 2)
}
\author{
Eric Hare \email{erichare@iastate.edu}
}

