% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{margins}
\alias{margins}
\title{Marginal distributions of a joint random variable}
\usage{
margins(X, sep = ",")
}
\arguments{
\item{X}{a random variable}

\item{sep}{parameter specifying the separator between dimensions, defaults to ","}
}
\description{
Extracts the marginal probability mass functions from a joint distribution.
}
\examples{
X <- RV(1:6, 1/6)
X3 <- iid(X, 3)
margins(X3)
}
\author{
Heike Hofmann \email{hofmann@iastate.edu}
}

