% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{plot.RV}
\alias{plot.RV}
\title{Plot a random variable of class "RV"}
\usage{
\method{plot}{RV}(x, ..., tol = 1e-10, pch = 16, cex = 1.2, lwd = 2,
  col = "black", xlab = "Possible Values", ylab = "Probabilities")
}
\arguments{
\item{x}{A random variable}

\item{...}{Additional arguments to be passed to the "plot" function}

\item{tol}{Only display outcomes with probabilities above tol}

\item{pch}{Either an integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default.}

\item{lwd}{The line width, a positive number, defaulting to 2.}

\item{col}{A specification for the default plotting color}

\item{xlab}{Label for the X axis}

\item{ylab}{Label for the Y axis}
}
\description{
Plot a random variable of class "RV"
}
\examples{
fair.die <- RV(1:6, rep(1/6,6))
plot(fair.die)
}

