% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.RV}
\alias{print.RV}
\title{Print a random variable of class "RV"}
\usage{
\method{print}{RV}(x, odds = attr(x, "odds"), fractions = attr(x,
  "fractions"), all.outcomes = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{A random variable}

\item{odds}{If TRUE, print as odds instead of probs}

\item{fractions}{If TRUE, print probs as fractions instead of decimals}

\item{all.outcomes}{If TRUE, print all outcomes rather than the first ten}

\item{digits}{Number of digits to print for probabilities}

\item{...}{Additional arguments to be passed to the "format" function}
}
\description{
Print a random variable of class "RV"
}
\examples{
fair.die <- make.RV(1:6, rep(1/6,6))
print(fair.die)
}
\author{
Eric Hare \email{erichare@iastate.edu}
}

