% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/discreteRV-Probabilities.r
\name{SofI}
\alias{SofI}
\title{Sum of independent random variables}
\usage{
SofI(..., fractions = attr(list(...)[[1]], "fractions"))
}
\arguments{
\item{...}{Arbitrary number of random variables}

\item{fractions}{If TRUE, return the probabilities as fractions}
}
\description{
Sum of independent random variables
}
\examples{
X.Bern <- RV(c(1,0), c(.5,.5))
X.fair.die <- RV(1:6, rep(1/6,6))

S5 <- SofI(X.Bern, X.Bern, X.Bern, X.Bern, X.Bern)
S.mix <- SofI(X.Bern, X.fair.die)  # Independent but not IID
}

